pipeline/requirements.txt
PyYAML>=6.0
matplotlib>=3.7.0
pytest>=7.0.0
